# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Hello World App
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths 
basePath: /
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
  - text/plain
# format of the responses to the client (Accepts)
produces:
  - application/json
x-a127-config: {}
x-a127-services: {}
paths:
  /hello:
    # binds a127 app logic to a route
    x-swagger-router-controller: hello_world
    x-a127-apply: {}
    get:
      description: Returns 'Hello' to the caller
      # used as the method name of the controller
      operationId: fuck
      parameters:
        - name: name
          in: query
          description: The name of the person to whom to say hello
          required: false
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
            
  /movie:
    # binds a127 app logic to a route
    x-swagger-router-controller: movie
    x-a127-apply: {}

    get:
      description: Returns requested movie records
      # used as the method name of the controller
      operationId: getMovie
      parameters:
        - name: reviews
          in: query
          description: The query used to delete specified records
          required: true
          type: boolean
        - name: ql
          in: query
          description: The query used to select records from the db
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

    post:
      description: Returns requested movie records
      # used as the method name of the controller
      operationId: postMovie

      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

    delete:
          description: deletes requested movie records
          # used as the method name of the controller
          operationId: deleteMovie
          parameters:
          - name: ql
            in: query
            description: The query used to delete specified records
            required: true
            type: string

          responses:
            "200":
              description: Success
              schema:
                # a pointer to a definition
                $ref: "#/definitions/HelloWorldResponse"
            # responses may fall through to errors
            default:
              description: Error
              schema:
                $ref: "#/definitions/ErrorResponse"
# complex objects have schema definitions
definitions:
  HelloWorldResponse:
    type: string
  ErrorResponse:
    type: string